// 00-reforzamiento/src/interfaces/reqRes.tsx
// Generated by https://quicktype.io

export interface ReqResListadoInterface {
    page:        number;
    per_page:    number;
    total:       number;
    total_pages: number;
    data:        UsuarioInterface[];
    support:     Support;
}

export interface UsuarioInterface {
    id:         number;
    email:      string;
    first_name: string;
    last_name:  string;
    avatar:     string;
}

export interface Support {
    url:  string;
    text: string;
}